[rename/#63] auth 패키지명 변경
