+++ app/src/main/java/com/bongtu/baekseo/data/datasource/auth/AuthDataSource.kt
@@ -1,11 +1,16 @@
 package com.bongtu.baekseo.data.datasource.auth
 
 import com.bongtu.baekseo.core.network.model.BaseResponse
-import com.bongtu.baekseo.data.dto.oauth.PostKakaoLoginRequest
-import com.bongtu.baekseo.data.dto.oauth.PostKakaoLoginResponse
+import com.bongtu.baekseo.data.dto.auth.PostKakaoLoginRequest
+import com.bongtu.baekseo.data.dto.auth.PostKakaoLoginResponse
+import com.bongtu.baekseo.data.dto.auth.PostSignUpRequest
 
 interface AuthDataSource {
     suspend fun postKakaoLogin(
         request: PostKakaoLoginRequest,
+    ): BaseResponse<PostKakaoLoginResponse>
+
+    suspend fun postSignUp(
+        request: PostSignUpRequest,
     ): BaseResponse<PostKakaoLoginResponse>
 }
