+++ app/src/main/java/com/bongtu/baekseo/data/datasourceimpl/auth/AuthDataSourceImpl.kt
@@ -2,8 +2,9 @@ package com.bongtu.baekseo.data.datasourceimpl.auth
 
 import com.bongtu.baekseo.core.network.model.BaseResponse
 import com.bongtu.baekseo.data.datasource.auth.AuthDataSource
-import com.bongtu.baekseo.data.dto.oauth.PostKakaoLoginRequest
-import com.bongtu.baekseo.data.dto.oauth.PostKakaoLoginResponse
+import com.bongtu.baekseo.data.dto.auth.PostKakaoLoginRequest
+import com.bongtu.baekseo.data.dto.auth.PostKakaoLoginResponse
+import com.bongtu.baekseo.data.dto.auth.PostSignUpRequest
 import com.bongtu.baekseo.data.service.auth.AuthService
 import javax.inject.Inject
 
@@ -12,4 +13,7 @@ class AuthDataSourceImpl @Inject constructor(
 ) : AuthDataSource {
     override suspend fun postKakaoLogin(request: PostKakaoLoginRequest): BaseResponse<PostKakaoLoginResponse> =
         authService.postKakaoLogin(request)
+
+    override suspend fun postSignUp(request: PostSignUpRequest): BaseResponse<PostKakaoLoginResponse> =
+        authService.postSignUp(request)
 }
