+++ app/src/main/java/com/bongtu/baekseo/data/mapper/AuthMapper.kt
@@ -1,6 +1,6 @@
 package com.bongtu.baekseo.data.mapper
 
-import com.bongtu.baekseo.data.dto.oauth.PostKakaoLoginResponse
+import com.bongtu.baekseo.data.dto.auth.PostKakaoLoginResponse
 import com.bongtu.baekseo.data.model.auth.KakaoLogin
 
 fun PostKakaoLoginResponse.toModel() = KakaoLogin(
